Upload account image

+++ src/components/EmployePosts.js
@@ -1,12 +1,20 @@
 import React, { useEffect, useState } from 'react'
 import axios from "axios";
 import useAuth from '../auth/useAuth';
+import { Modal, Button } from "react-bootstrap";
+import "../styles/EmployeePosts.css"
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
+
+const MySwal = withReactContent(Swal)
 
 const EmployePosts = (props) => {
     const { user } = useAuth();
     const [userSession, setUserSession] = useState({})
     const [userReceiver, setUserReceiver] = useState(null)
     const [posts, setPosts] = useState([])
+    const [show, setShow] = useState(false)
+    const [description, setDescription] = useState(``)
 
 
     useEffect(() => {
@@ -30,15 +38,83 @@ const EmployePosts = (props) => {
     }
 
     let button;
-    let h1;
     if (userSession.id !== parseInt(props.match.params.id)) {
         button = <button className="btn btn-primary">Dar una estrella</button>
     }
 
     const deletePost = async (post) => {
-        await axios.delete(`http://localhost:3001/posts/${post.id}`)
-        getPosts();
+        try {
+            MySwal.fire({
+                title: `¿Seguro quieres eliminar este post?`,
+                showDenyButton: true,
+                showCancelButton: true,
+                showConfirmButton: false,
+                denyButtonText: `Eliminar`,
+                cancelButtonText: `Cancelar`,
+                icon: 'warning',
+            }).then(async (result) => {
+                /* Read more about isConfirmed, isDenied below */
+                if (result.isDenied) {
+                    await axios.delete(`http://localhost:3001/posts/${post.id}`)
+                    MySwal.fire({
+                        icon: 'success',
+                        title: 'Proyecto eliminado!',
+                        showConfirmButton: false,
+                        timer: 1000
+                    })
+                    getPosts();
+                }
+            })
+
+
+        } catch (error) {
+            MySwal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salió mal!',
+                footer: `<p>${error.name}: ${error.message}</p>`
+            })
+        }
     }
+
+    const editPost = async (post) => {
+        if (description === ``) {
+            MySwal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Parece que no escribiste nada'
+            })
+        } else {
+            try {
+                posts[post.id - 1].post = description
+                await axios.put(`http://localhost:3001/posts/${post.id}`, post)
+                MySwal.fire({
+                    icon: 'success',
+                    title: 'Post editado correctamente!',
+                    showConfirmButton: false,
+                    timer: 1000
+                })
+                handleClose();
+                getPosts();
+            } catch (error) {
+                MySwal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Algo salió mal!',
+                    footer: `<p>${error.name}: ${error.message}</p>`
+                })
+                handleClose();
+            }
+        }
+    }
+
+    const onInputChange = (e) => {
+        setDescription(e.target.value)
+    }
+
+    const handleClose = () => {
+        setShow(!show);
+    };
 
     return (
         <div>
@@ -67,11 +143,32 @@ const EmployePosts = (props) => {
                                                     {
                                                         post.author.id === userSession.id &&
                                                         <div>
-                                                            <button className="btn btn-outline-danger me-2" onClick={() => deletePost(post)}>Eliminar post</button>
-                                                            <button className="btn btn-outline-primary" onClick={() => deletePost(post)}>Editar post</button>
+                                                            <button className="btn btn-outline-primary me-2" onClick={() => setShow(true)}>Editar post</button>
+                                                            <button className="btn btn-outline-danger" onClick={() => deletePost(post)}>Eliminar post</button>
                                                         </div>
                                                     }
                                                 </div>
+                                                <Modal show={show} onHide={handleClose}>
+                                                    <Modal.Header closeButton>
+                                                        <Modal.Title>Edita la descripción de tu post</Modal.Title>
+                                                    </Modal.Header>
+                                                    <Modal.Body>
+                                                        <div className="row justify-content-center">
+                                                            <textarea className="inp px-4 pt-4" style={{
+                                                                height: "10em",
+                                                                width: "80%"
+                                                            }} onChange={onInputChange} type="text" minLength="1" rows="5" cols="40" maxLength="500" name="name" placeholder="Descripción del post" />
+                                                        </div>
+                                                    </Modal.Body>
+                                                    <Modal.Footer>
+                                                        <Button variant="primary" onClick={() => editPost(post)}>
+                                                            Editar
+                                                        </Button>
+                                                        <Button variant="secondary" onClick={handleClose}>
+                                                            Cerrar
+                                                        </Button>
+                                                    </Modal.Footer>
+                                                </Modal>
                                             </div>
                                             <div className="card-text">
                                                 <p className="m-0 ms-2">{post.post}</p>
